Move from PyCrypto to pycryptodomex
Streamlink depend on pycryptodome python module, but that module is not
available in Debian. PyCrypto support that was used before is not
working anymore because of new hard dependencies on pycryptodome instead of
PyCrypto.

So use pycryptodomex instead using a patch to rename 'from Crypto.' to
'from Cryptodome.' and change the used package.